Revert "Merge remote-tracking branch 'origin/master' into clkng"
This reverts commit 949e46183d86f623822fc7dc4173c0d580fc1adb, reversing
changes made to dfa9d117a290eb11efd573fdf16f3709f720a829.

+++ contrib/bsnmp/snmpd/config.c
@@ -663,7 +663,7 @@ gettoken()
 			printf("(EOL)");
 			break;
 		  case TOK_NUM:
-			printf("(NUM %ju)", (uintmax_t)numval);
+			printf("(NUM %llu)", numval);
 			break;
 		  case TOK_STR:
 			printf("(STR %.*s)", (int)strvallen, strval);
